moved some state to Dashboard.js for future functionality
