+++ src/components/timeline/timeline.js
@@ -1,33 +1,8 @@
 import React from 'react';
 import { Timeline, TimelineItem } from 'vertical-timeline-component-for-react';
 import Moment from 'react-moment';
-import axios from 'axios';
-import { useState, useEffect } from 'react';
 
-function latest(array, n) {
-  return array.slice(0, n);
-}
-
-export default function PbTimeline() {
-  const [results, setResults] = useState([]);
-
-  useEffect(() => {
-    async function getData() {
-      await axios
-        .get(`https://api.846policebrutality.com/api/incidents`)
-        .then(res => {
-          setResults(latest(res.data.data, 10));
-        })
-        .catch(err => {
-          console.log(err);
-        });
-    }
-
-    getData();
-  }, []);
-
-  console.log(results);
-
+export default function PbTimeline({ results, setSearch }) {
   return (
     <>
       <div className="tl-header">
